+++ hinatablogimg/logconfig.yml
@@ -1,7 +1,7 @@
 version: 1
 formatters:
   simple:
-    format: '%(asctime)s [%(name)s]:(%(lineno)d) %(message)s '
+    format: '%(asctime)s [%(levelname)s]:(%(lineno)d) %(message)s '
 handlers:
   filehandler:
     class: logging.handlers.RotatingFileHandler
